Update Demo to be compliant to v2 Demo
- Added new CSS Class to allow overlay of the `div.wizard-card` See
demo; is used for the chosen components
- Updated CSS of wizard steps navigation, when more than 4 steps the
progress bar was not pushed down to the bottom. (Fixed)
- Added Chosen to demo v3
- Added images from v2 demo
- Removed no-modal code for buttons
- Added dimension code, wizard steps height was incorrect, body height
is now also applied to wizard-steps to set the correct height when there
are more than 4 steps as was in the first / static build
- Updated de demo.html, incorperated all the card from demo v2,
refactored code to boostrap v3, reuqired for buttons / input /
form-controls, maintained first demo setup of first card to show error
popover with input field with adjacent button.